Improved to zip up the build once the UI's html task is complete
